Revert "refactor: Use meaningful package names"
